# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Pheromone Management API
  version: "0.1.0"
  description: |
    REST API for the Pheromone digital-twin platform management server.

    All endpoints under `/api/v1/` (except `/api/v1/auth/login` and
    `/api/v1/health`) require a Bearer token obtained via
    `POST /api/v1/auth/login`.

    Real-time agent and twin state updates are available via the
    Server-Sent Events stream at `GET /api/v1/events`.

    Kubernetes liveness and readiness probes are at `/healthz` and
    `/readyz` respectively (no authentication required).
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: Authenticate and obtain a Bearer token
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: HMAC-SHA256-signed token obtained from POST /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    MethodNotAllowed:
      description: HTTP method not supported for this endpoint
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [token, expires_at, user]
      properties:
        token:
          type: string
          description: Bearer token for subsequent requests
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, viewer]
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: HTTP status text
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Human-readable error detail